requirements.txt
pydantic==2.6.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-arango==7.9.1